Initial attempt.
