+++ benchmark/src/lib.rs
@@ -1,5 +1,5 @@
 pub use async_graphql::http::GQLResponse;
-use async_graphql::{ObjectType, QueryResponse, Schema, SubscriptionType};
+use async_graphql::{GQLQueryResponse, ObjectType, Schema, SubscriptionType};
 use async_graphql_parser::{parse_query, types::ExecutableDocument};
 use async_std::task;
 
@@ -10,7 +10,7 @@ static ALLOC: jemallocator::Jemalloc = jemallocator::Jemalloc;
 pub fn run<Query, Mutation, Subscription>(
     s: &Schema<Query, Mutation, Subscription>,
     q: &str,
-) -> QueryResponse
+) -> GQLQueryResponse
 where
     Query: ObjectType + Send + Sync + 'static,
     Mutation: ObjectType + Send + Sync + 'static,
