+++ integrations/rocket/src/lib.rs
@@ -4,7 +4,7 @@
 #![forbid(unsafe_code)]
 
 use async_graphql::{
-    IntoQueryBuilder, IntoQueryBuilderOpts, ObjectType, QueryBuilder, QueryResponse, Schema,
+    GQLQueryResponse, IntoQueryBuilder, ObjectType, QueryBuilder, ReceiveMultipartOptions, Schema,
     SubscriptionType, Variables,
 };
 use log::{error, info};
@@ -99,7 +99,7 @@ impl GraphQL {
     /// ```
     pub fn fairing_with_opts<Q, M, S>(
         schema: Schema<Q, M, S>,
-        opts: IntoQueryBuilderOpts,
+        opts: ReceiveMultipartOptions,
     ) -> impl Fairing
     where
         Q: ObjectType + Send + Sync + 'static,
@@ -109,7 +109,7 @@ impl GraphQL {
         GraphQL::attach(schema, opts)
     }
 
-    fn attach<Q, M, S>(schema: Schema<Q, M, S>, opts: IntoQueryBuilderOpts) -> impl Fairing
+    fn attach<Q, M, S>(schema: Schema<Q, M, S>, opts: ReceiveMultipartOptions) -> impl Fairing
     where
         Q: ObjectType + Send + Sync + 'static,
         M: ObjectType + Send + Sync + 'static,
@@ -231,7 +231,7 @@ impl FromData for GQLRequest {
     type Error = String;
 
     async fn from_data(req: &Request<'_>, data: Data) -> data::Outcome<Self, Self::Error> {
-        let opts = match req.guard::<State<'_, Arc<IntoQueryBuilderOpts>>>().await {
+        let opts = match req.guard::<State<'_, Arc<ReceiveMultipartOptions>>>().await {
             Outcome::Success(opts) => opts,
             Outcome::Failure(_) => {
                 return data::Outcome::Failure((
@@ -258,7 +258,7 @@ impl FromData for GQLRequest {
 /// Wrapper around `async-graphql::query::QueryResponse` for implementing the trait
 /// `rocket::response::responder::Responder`, so that `GQLResponse` can directly be returned
 /// from a Rocket Route function.
-pub struct GQLResponse(pub QueryResponse);
+pub struct GQLResponse(pub GQLQueryResponse);
 
 impl<'r> Responder<'r, 'static> for GQLResponse {
     fn respond_to(self, _: &'r Request<'_>) -> response::Result<'static> {
@@ -277,13 +277,13 @@ impl<'r> Responder<'r, 'static> for GQLResponse {
 /// Extension trait, to allow the use of `cache_control` with for example `ResponseBuilder`.
 pub trait CacheControl {
     /// Add the `async-graphql::query::QueryResponse` cache control value as header to the Rocket response.
-    fn cache_control(&mut self, resp: &async_graphql::Result<QueryResponse>) -> &mut Self;
+    fn cache_control(&mut self, resp: &async_graphql::Result<GQLQueryResponse>) -> &mut Self;
 }
 
 impl<'r> CacheControl for ResponseBuilder<'r> {
     fn cache_control(
         &mut self,
-        resp: &async_graphql::Result<QueryResponse>,
+        resp: &async_graphql::Result<GQLQueryResponse>,
     ) -> &mut ResponseBuilder<'r> {
         match resp {
             Ok(resp) if resp.cache_control.value().is_some() => self.header(Header::new(
